These includes all the changes from Ethan Main
The app.yaml is the same one we've all been using.  The main.py is the file we covered on the call today with all the changes Ethan made as well as the changes I made and Eric helped us make on the call.  The sendPost.html is the file Ethan made with a change I made to stringify the JSON and also to include the dictionary changes Eric suggested on the call.